app/requirements.txt
fastapi[all]
uvicorn[standard]
psycopg2-binary
jinja2